src/soli-localstack/soliBackend/requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14.0
gunicorn>=21.0.0
boto3>=1.28.0
localstack-client>=1.39
psycopg2-binary>=2.9.0 
django-cors-headers>=4.3.0
djangorestframework-simplejwt
